feature/GH-2419: in routing module, move state into store folder
