+++ projects/core/src/routing/store/reducers/router.reducer.ts
@@ -11,7 +11,7 @@ import {
 import { PageType } from '../../../model/cms.model';
 import { CmsActivatedRouteSnapshot } from '../../models/cms-route';
 import { PageContext } from '../../models/page-context.model';
-import { ROUTING_FEATURE } from '../../state';
+import { ROUTING_FEATURE } from '../state';
 import * as fromActions from '../actions';
 
 export interface RouterState
